+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-## 1.x.x [2022/XX/XX]
+## 1.7.0 [2022/XX/XX]
 
 **New Features:**
 
@@ -8,6 +8,7 @@
 
 - Changed Event ID Statistics option to Event ID Metrics option. (`-s, --statistics`  -> `-M, --metrics`) (#706) (@hitenkoku)
   (Note: `statistics_event_info.txt` was changed to `event_id_info.txt`.)
+- Added display new version of Hayabusa when updating. (#710) (@hitenkoku)
 
 **Bug Fixes:**